refactor(engine): move MockHBaseStorageBackend to hbase package
More logical package location since it's HBase-specific.
