+++ engine/src/main/kotlin/com/kakao/actionbase/v2/engine/storage/hbase/MockHBaseStorageBackend.kt
@@ -1,11 +1,7 @@
-package com.kakao.actionbase.v2.engine.storage.mock
+package com.kakao.actionbase.v2.engine.storage.hbase
 
 import com.kakao.actionbase.v2.engine.storage.StorageBackend
 import com.kakao.actionbase.v2.engine.storage.StorageBuckets
-import com.kakao.actionbase.v2.engine.storage.hbase.HBaseConnections
-import com.kakao.actionbase.v2.engine.storage.hbase.HBaseStorageBucket
-import com.kakao.actionbase.v2.engine.storage.hbase.HBaseTable
-import com.kakao.actionbase.v2.engine.storage.hbase.HBaseTables
 import com.kakao.actionbase.v2.engine.storage.hbase.impl.NewMockTable
 
 import org.apache.hadoop.hbase.TableName
