+++ engine/src/test/kotlin/com/kakao/actionbase/v2/engine/storage/DefaultStorageBackendFactoryTest.kt
@@ -1,7 +1,7 @@
 package com.kakao.actionbase.v2.engine.storage
 
+import com.kakao.actionbase.v2.engine.storage.hbase.MockHBaseStorageBackend
 import com.kakao.actionbase.v2.engine.storage.memory.MemoryStorageBackend
-import com.kakao.actionbase.v2.engine.storage.mock.MockHBaseStorageBackend
 
 import org.junit.jupiter.api.AfterEach
 import org.junit.jupiter.api.DisplayName
